Move components to their own files and other improvements
