+++ src/common/components/deleteButton/deleteButton.module.scss
@@ -1,0 +1,19 @@
+@import 'colors.scss';
+
+.deleteIconWrapper {
+  width: 1rem;
+  margin-right: 1rem;
+  fill: red;
+
+  svg {
+    display: block;
+  }
+}
+
+.deleteImageButton {
+  margin: auto;
+  display: flex;
+  margin-top: 42px;
+  align-items: center;
+  color: $hel-bus;
+}
